Ready to print.
it is what it is. Whoever is reading this in the future, I really did my best. It is feature-complete. The science is good. The writing is not my best, as I ran out of time. I wish I could have made it better. But it is what it is, and what it is is a finished dissertation.